review: round-18 fix — move isAllInterfacesHost to core/node, wildcard→localhost in CLI deep-link (#1795)
Address AC1 of the round-18 review of #1796 (converged; AC2/AC3 acknowledged):

- AC1: the CLI handoff deep-link canonicalized the host (round 15) but didn't do
  the wildcard→localhost substitution the banner/sandbox URL do, so HOST=0.0.0.0
  gave an awkward http://0.0.0.0:PORT deep link (allow-listed, so it connects,
  but not a nice URL to hand a human). Moved isAllInterfacesHost + its pure
  helpers (canonicalizeIpv6, isAllZeroIpv4, parseAddressPart,
  ALL_INTERFACES_LITERALS) from clients/web/server/resolve-bind-host.ts into
  core/node/hostUrl.ts — the correct home for a pure host predicate, alongside
  canonicalUrlHost (same consolidation canonicalUrlHost got in round 14) — so the
  CLI can import it. buildHandoff now does isAllInterfacesHost(host) ? "localhost"
  : canonicalUrlHost(host). All four consumers (guard, banner, sandbox, CLI)
  share one predicate. Moved the isAllInterfacesHost tests to hostUrl.test.ts;
  added a CLI test asserting the wildcard deep link is http://localhost:PORT.

AC2 (CLI reads CLIENT_PORT raw): left as-is — the same env fails the web server
outright (buildWebServerConfig throws), so a handoff pointing at it is moot.
AC3 (inert start()-twice NaN + broader CSP fallback): pre-existing/inert,
unchanged.
